Fixed compiler errors from -Wall with icpc 10.1. Updated findstars parameters to work with some files that were failing. Parallelized findstars code. Got the transformation class to work with the DES WCS parameters.   -- Note, the definitions of PV1 and PV2 are backwards from what I expected.      Either that or the SExtractor catalog ra and dec are wrong....  Hmm. Added ra and dec to SXCat class. Fixed the extensions in wl.config to be .fits for those that are now FITS. Added option to SConstruct to compile with static linkage. Added path options to SConstruct as an alternate way to add library/include     paths other than IMPORT_PATHS. Changed int32's to long.  int32 was clashing with another int32 in some     header file for my icpc 10.1 compilation.  I didn't bother tracking it     down.  It was easier to just change them to long.  For what we do with     them, this is portable enough. Fixed a bug in the FittedPSF FITS I/O with rot_matrix. Cleaned up the interface for FittedPSF I/O -- it is now closer to a form     that will facilitate the kind of generalized I/O capability that     we want, rather than hardcoded for DES format. Moved stuff around in wl.config.  I added comments that refer to the     way I think we want the code to work, even though some of it isn't     implemented yet.  These parts are marked by ***. Also, changed the names of some of the parameters to make everything more     consistent.  Mostly involved adding prefixes to many parameter names.
git-svn-id: https://desweb.cosmology.illinois.edu/svn/desdm/devel/wl/trunk@2673 2422e80f-9519-0410-a819-8e21e7c86b6d

+++ src/Pixel.cpp
@@ -5,7 +5,7 @@
 void GetPixList(const Image<double>& im, std::vector<Pixel>& pix,
     const Position cen, double sky, double noise, double gain,
     const Image<double>* wt_im, const Transformation& trans,
-    double aperture, int& flag)
+    double aperture, long& flag)
 {
   xdbg<<"Start GetPixList\n";
   if (wt_im) {
